requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt etc/requirements.in
#
aniso8601==7.0.0          # via graphene
django-cors-headers==3.1.0
django==2.2
djangorestframework==3.10.3
graphene-django==2.5.0
graphene==2.1.8           # via graphene-django
graphql-core==2.2.1       # via graphene, graphene-django, graphql-relay
graphql-relay==2.0.0      # via graphene
pillow==6.1.0
promise==2.2.1            # via graphene-django, graphql-core, graphql-relay
pytz==2019.2              # via django
rx==1.6.1                 # via graphql-core
singledispatch==3.4.0.3   # via graphene-django
six==1.12.0               # via graphene, graphene-django, graphql-core, graphql-relay, promise, singledispatch
sqlparse==0.3.0           # via django
